Move out consensus logic from ToC
