+++ src/tonic/mod.rs
@@ -14,7 +14,7 @@ use ::api::grpc::qdrant::{HealthCheckReply, HealthCheckRequest};
 use std::net::{IpAddr, SocketAddr};
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
-use storage::content_manager::toc::TableOfContent;
+use storage::Dispatcher;
 use tokio::{runtime, signal};
 use tonic::{transport::Server, Request, Response, Status};
 
@@ -31,7 +31,7 @@ impl Qdrant for QdrantService {
     }
 }
 
-pub fn init(toc: Arc<TableOfContent>, host: String, grpc_port: u16) -> std::io::Result<()> {
+pub fn init(dispatcher: Arc<Dispatcher>, host: String, grpc_port: u16) -> std::io::Result<()> {
     let tonic_runtime = runtime::Builder::new_multi_thread()
         .enable_io()
         .enable_time()
@@ -46,8 +46,8 @@ pub fn init(toc: Arc<TableOfContent>, host: String, grpc_port: u16) -> std::io::
             let socket = SocketAddr::from((host.parse::<IpAddr>().unwrap(), grpc_port));
 
             let service = QdrantService::default();
-            let collections_service = CollectionsService::new(toc.clone());
-            let points_service = PointsService::new(toc.clone());
+            let collections_service = CollectionsService::new(dispatcher.clone());
+            let points_service = PointsService::new(dispatcher.toc().clone());
 
             log::info!("Qdrant gRPC listening on {}", grpc_port);
 
@@ -66,7 +66,7 @@ pub fn init(toc: Arc<TableOfContent>, host: String, grpc_port: u16) -> std::io::
 }
 
 pub fn init_internal(
-    toc: Arc<TableOfContent>,
+    dispatcher: Arc<Dispatcher>,
     host: String,
     internal_grpc_port: u16,
     to_consensus: std::sync::mpsc::SyncSender<crate::consensus::Message>,
@@ -74,6 +74,7 @@ pub fn init_internal(
     use crate::tonic::api::raft_api::RaftService;
     use ::api::grpc::qdrant::raft_server::RaftServer;
 
+    let toc = dispatcher.toc().clone();
     let tonic_runtime = runtime::Builder::new_multi_thread()
         .enable_io()
         .enable_time()
@@ -90,7 +91,7 @@ pub fn init_internal(
             let service = QdrantService::default();
             let collections_internal_service = CollectionsInternalService::new(toc.clone());
             let points_internal_service = PointsInternalService::new(toc.clone());
-            let raft_service = RaftService::new(to_consensus, toc.clone());
+            let raft_service = RaftService::new(to_consensus, dispatcher.clone());
 
             log::debug!("Qdrant internal gRPC listening on {}", internal_grpc_port);
