+++ tests/consensus_tests/test_collection_created_before.py
@@ -18,7 +18,7 @@ def test_collection_before_peers_added(tmp_path: pathlib.Path):
     peer_api_uris.append(bootstrap_api_uri)
 
     # Wait for leader
-    wait_for_leader_setup(bootstrap_api_uri)
+    wait_peer_added(bootstrap_api_uri)
 
     # Create collection
     r = requests.put(
@@ -34,7 +34,7 @@ def test_collection_before_peers_added(tmp_path: pathlib.Path):
         peer_api_uris.append(start_peer(
             peer_dirs[i], f"peer_0_{i}.log", bootstrap_uri))
         # Add peers one by one sequentially
-        wait_for_leader_setup(peer_api_uris[i])
+        wait_peer_added(peer_api_uris[i])
 
     # Wait for cluster
     wait_for_uniform_cluster_size(peer_api_uris)
